// apps/web/src/api/generated/Stats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProjectCount } from "./ProjectCount";
import type { TagCount } from "./TagCount";
import type { TopicCount } from "./TopicCount";
import type { WeekCount } from "./WeekCount";

export type Stats = { notes: number, topics: number, repeated_topics: number, annotations: number, top_topics: Array<TopicCount>, tags: Array<TagCount>, by_week: Array<WeekCount>, by_project: Array<ProjectCount>, conflicts: number, invalid_files: number, };